global slot removal, symbol refactors
