+++ src/index.tsx
@@ -3,6 +3,7 @@ import * as React from 'react';
 import ConditionalSlot, {
   IConditionalSlot, IConditionalSlotBase, createConditionalSlot,
 } from './ConditionalSlot';
+import useChildren from './utils/useChildren';
 
 interface ISlot<T> {
   /**
@@ -211,49 +212,8 @@ export const createSlot: IOverloadCreateSlot = <T extends {} = {}, S extends {} 
     return SlottedElement;
 };
 
-/**
- * Indexes React children for faster access by Slot components
- * @param scope - react children, in any format
- */
-export const useChildren = (scope: any): IIndexedChildren => {
-  if (scope === undefined || scope.length === 0) {
-    return new Map();
-  }
-  const childrenCount = React.Children.count(scope);
-  const result = new Map<symbol | string, ISortChildrenEl[]>();
-  const injectSlot = (child: JSX.Element, index: number) => {
-    let childType = 'string';
-    if (React.isValidElement(child) && child.type !== undefined) {
-      const obj: any = child.type;
-      if (obj.hasOwnProperty('displayName')) {
-        childType = obj.displayName || 'string';
-      } else {
-        childType = obj;
-      }
-      if (obj.hasOwnProperty('displaySymbol')) {
-        childType = obj.displaySymbol;
-      }
-    }
-    const resultGet = result.get(childType);
-    if (resultGet === undefined) {
-      result.set(childType, [{index, child}]);
-    } else {
-      resultGet.push({index, child});
-      result.set(childType, resultGet);
-    }
-  };
-  if (childrenCount === 1) {
-    injectSlot(scope, 0);
-  } else if (childrenCount > 1) {
-      React.Children.forEach(scope, injectSlot);
-  }
-  result.set('$$isSlottedChildren', []);
-  return result;
-};
-
 import FilterSlot from './FilterSlot';
 export { FilterSlot as NonSlotted };
-export { FilterSlot };
-export { ConditionalSlot };
+export { FilterSlot as FilterSlot, useChildren, ConditionalSlot };
 
 export default createSlot;
